Module chomp::ascii

Utilities and parsers for dealing with ASCII data in u8 format.

Traits

Float
Trait enabling the conversion from a matched Buffer to a float of the correct type.

Functions

decimal
Parses a series of digits and converts them to an integer.

digit
Parses a single digit.

float
Parses a float into a f64 or f32, will error with an Error::unexpected if the float does not map to a proper float.

is_alpha
ASCII alphabetic predicate.

is_alphanumeric
ASCII alphanumeric predicate.

is_digit
ASCII digit predicate.

is_end_of_line
A predicate matching eithr a newline ('\n') or a carriage return ('\r').

is_horizontal_space
A predicate which matches either space (' ') or horizontal tab ('\t').

is_lowercase
Lowercase ASCII predicate.

is_uppercase
Uppercase ASCII character predicate.

is_whitespace
ASCII whitespace predicate.

match_float
Matches a floating point number in base-10 with an optional exponent, returning a buffer.

signed
Parses a number with an optional leading '+' or '-'.

skip_whitespace
Skips over whitespace.

string_ci
Matches the given slice against the parser in a case-insensitive manner, returning the matched slice upon success. Only respects ASCII characters for the case-insensitive comparison.