Module chomp::buffer::data_source

Implementation of datasources for Source.

Structs

IteratorDataSource
Implementation of DataSource for Iterator.

RWDataSource
Implementation of DataSource for streams (e.g. network connections) that you can Read and Write. It's really helpful to have the ability to write through the DataSource and Source objects, because once created, often they take full ownership of your input stream, and when working with bidirectional connections you still need a way to write to them.

ReadDataSource
Implementation of DataSource for io::Read instances.

Traits

DataSource
Abstraction over io::Read, Iterator and others.