Struct chomp::buffer::data_source::RWDataSource

pub struct RWDataSource<RW: Read + Write>(_);

Implementation of DataSource for streams (e.g. network connections) that you can Read and Write. It's really helpful to have the ability to write through the DataSource and Source objects, because once created, often they take full ownership of your input stream, and when working with bidirectional connections you still need a way to write to them.

Methods

impl<RW: Read + Write> RWDataSource<RW>

fn new(inner: RW) -> Self

Creates a new RWDataSource from a stream (e.g. network connection).

fn into_inner(self) -> RW

Consumes self to reveal the underlying stream.