Enum chomp::buffer::StreamError

pub enum StreamError<'a, I, E> where I: 'a {
    ParseError(&'a [I], E),
    Incomplete(usize),
    IoError(Error),
    EndOfInput,
    Retry,
}

Error type for parsing using the Stream trait.

Variants

ParseError

An error occurred in the parser, the given slice indicates the part which failed.

Incomplete

Parser failed to complete with the available data.

IoError

An IO-error occurred while attempting to fill the buffer.

EndOfInput

The last parser completed successfully and there is no more input to parse.

Retry

The last parser failed with an incomplete state, fill the buffer and try again.

Filling the buffer is automatic by default.