Struct chomp::buffer::InputBuf

#[must_use]

pub struct InputBuf<'a, I: 'a>(_, _);

Input buffer type which contains a flag which tells if we might need to read more data.

Methods

impl<'a, I: 'a> InputBuf<'a, I>

fn new(buf: &'a [I]) -> Self

Creates a new input buffer with incomplete set to false.

fn is_incomplete(&self) -> bool

Returns true if parsers want to obtain more data.

The result of the parsing is only accurate if this is false after completed parsing.

fn len(&self) -> usize

Returns the length of the contained buffer, may be an incomplete buffer.

fn is_empty(&self) -> bool

Returns true if the contained buffer is empty, may return true even when incomplete.