Struct chomp::buffer::Source

pub struct Source<S: DataSource, B: Buffer<S::Item>> { /* fields omitted */ }

Manages a buffer and data source pair, enabling efficient parsing from a streaming source.

Methods

impl<R: Read> Source<ReadDataSource<R>, FixedSizeBuffer<u8>>

fn new(source: R) -> Self

Creates a new Source from a Read instance with the default FixedSizeBuffer settings.

impl<R: Read, B: Buffer<u8>> Source<ReadDataSource<R>, B>

fn from_read(source: R, buffer: B) -> Self

Creates a new Source from Read and buffer instances.

impl<RW: Read + Write> Source<RWDataSource<RW>, FixedSizeBuffer<u8>>

fn new_rw(rwsource: RW) -> Self

Creates a new Source from Read+Write with the default FixedSizeBuffer settings.

impl<RW: Read + Write, B: Buffer<u8>> Source<RWDataSource<RW>, B>

fn from_read_write(source: RW, buffer: B) -> Self

Creates a new Source from Read+Write and buffer instances.

impl<I: Iterator, B: Buffer<I::Item>> Source<IteratorDataSource<I>, B> where
I::Item: Copy + PartialEq,

fn from_iter(source: I, buffer: B) -> Self

Creates a new Source from Iterator and Buffer instances.

impl<S: DataSource, B: Buffer<S::Item>> Source<S, B>

fn with_buffer(source: S, buffer: B) -> Self

Creates a new Source from DataSource and Buffer instances.

fn fill(&mut self) -> Result<usize>

Attempts to fill the buffer to satisfy the last call to parse().

fn len(&self) -> usize

Returns the number of bytes left in the buffer which have not yet been parsed.

fn is_empty(&self) -> bool

If the buffer is empty and the reader has reached the end.

fn capacity(&self) -> usize

Returns the capacity of the underlying buffer.

This is the maximum number of input items the buffer can store.

fn buffer(&self) -> &[S::Item]

Borrows the remainder of the buffer.

fn reset(&mut self)

Resets the buffer state, keeping the current buffer contents and cursor position.

This is useful when streaming data and more data has been made available on a socket/stream.

fn set_autofill(&mut self, value: bool)

Changes the setting automatic fill feature, true will make the buffer automatically call fill() on the next call to parse() after a Retry was encountered.