Function chomp::combinators::bounded::many

pub fn many<'a, I, T, E, F, U, R>(i: Input<'a, I>, r: R, f: F) -> ParseResult<'a, I, T, E> where I: Copy, R: BoundedRange, U: 'a, F: FnMut(Input<'a, I>) -> ParseResult<'a, I, U, E>, T: FromIterator<U>

Applies the parser F multiple times until it fails or the maximum value of the range has been reached, collecting the successful values into a T: FromIterator.

Propagates errors if the minimum number of iterations has not been met

Panics

Will panic if the end of the range is smaller than the start of the range.

Notes

Will allocate depending on the FromIterator implementation.

Will never yield more items than the upper bound of the range.

If the last parser succeeds on the last input item then this parser is still considered incomplete if the input flag END_OF_INPUT is not set as there might be more data to fill.