Function chomp::combinators::bounded::many_till

pub fn many_till<'a, I, T, E, R, F, U, N, P, V>(i: Input<'a, I>, r: R, p: P, end: F) -> ParseResult<'a, I, T, E> where I: Copy, U: 'a, V: 'a, N: 'a, R: BoundedRange, T: FromIterator<U>, P: FnMut(Input<'a, I>) -> ParseResult<'a, I, U, E>, F: FnMut(Input<'a, I>) -> ParseResult<'a, I, V, N>

Applies the parser P multiple times until the parser F succeeds and returns a value populated by the values yielded by P. Consumes the matched part of F. If F does not succeed within the given range R this combinator will propagate any failure from P.

Panics

Will panic if the end of the range is smaller than the start of the range.

Notes

Will allocate depending on the FromIterator implementation.

Will never yield more items than the upper bound of the range.

If the last parser succeeds on the last input item then this combinator is still considered incomplete unless the parser F matches or the lower bound has not been met.