Trait chomp::combinators::bounded::BoundedRange

pub trait BoundedRange {
    fn parse_many<'a, I, T, E, F, U>(self, Input<'a, I>, F) -> ParseResult<'a, I, T, E> where I: Copy, U: 'a, F: FnMut(Input<'a, I>) -> ParseResult<'a, I, U, E>, T: FromIterator<U>;
    fn skip_many<'a, I, T, E, F>(self, Input<'a, I>, F) -> ParseResult<'a, I, (), E> where T: 'a, F: FnMut(Input<'a, I>) -> ParseResult<'a, I, T, E>;
    fn many_till<'a, I, T, E, R, F, U, N, V>(self, i: Input<'a, I>, p: R, end: F) -> ParseResult<'a, I, T, E> where I: Copy, U: 'a, V: 'a, N: 'a, T: FromIterator<U>, R: FnMut(Input<'a, I>) -> ParseResult<'a, I, U, E>, F: FnMut(Input<'a, I>) -> ParseResult<'a, I, V, N>;
}

Trait for applying a parser multiple times based on a range.

Required Methods

fn parse_many<'a, I, T, E, F, U>(self, Input<'a, I>, F) -> ParseResult<'a, I, T, E> where I: Copy, U: 'a, F: FnMut(Input<'a, I>) -> ParseResult<'a, I, U, E>, T: FromIterator<U>

Applies the parser F multiple times until it fails or the maximum value of the range has been reached, collecting the successful values into a T: FromIterator.

Propagates errors if the minimum number of iterations has not been met

Panics

Will panic if the end of the range is smaller than the start of the range.

Notes

Will allocate depending on the FromIterator implementation.

Must never yield more items than the upper bound of the range.

Use combinators::bounded::many instead of calling this trait method directly.

If the last parser succeeds on the last input item then this parser is still considered incomplete if the input flag END_OF_INPUT is not set as there might be more data to fill.

fn skip_many<'a, I, T, E, F>(self, Input<'a, I>, F) -> ParseResult<'a, I, (), E> where T: 'a, F: FnMut(Input<'a, I>) -> ParseResult<'a, I, T, E>

Applies the parser F multiple times until it fails or the maximum value of the range has been reached, throwing away any produced value.

Propagates errors if the minimum number of iterations has not been met

Panics

Will panic if the end of the range is smaller than the start of the range.

Notes

Must never yield more items than the upper bound of the range.

Use combinators::bounded::many instead of calling this trait method directly.

If the last parser succeeds on the last input item then this parser is still considered incomplete if the input flag END_OF_INPUT is not set as there might be more data to fill.

fn many_till<'a, I, T, E, R, F, U, N, V>(self, i: Input<'a, I>, p: R, end: F) -> ParseResult<'a, I, T, E> where I: Copy, U: 'a, V: 'a, N: 'a, T: FromIterator<U>, R: FnMut(Input<'a, I>) -> ParseResult<'a, I, U, E>, F: FnMut(Input<'a, I>) -> ParseResult<'a, I, V, N>

Applies the parser P multiple times until the parser F succeeds and returns a value populated by the values yielded by P. Consumes the matched part of F. If F does not succeed within the given range R this combinator will propagate any failure from P.

Panics

Will panic if the end of the range is smaller than the start of the range.

Notes

Will allocate depending on the FromIterator implementation.

Use combinators::bounded::many_till instead of calling this trait method directly.

Must never yield more items than the upper bound of the range.

If the last parser succeeds on the last input item then this combinator is still considered incomplete unless the parser F matches or the lower bound has not been met.