Module chomp::prelude

Basic prelude.

Reexports

pub use parsers::any;

pub use parsers::eof;

pub use parsers::not_token;

pub use parsers::peek;

pub use parsers::peek_next;

pub use parsers::run_scanner;

pub use parsers::satisfy;

pub use parsers::satisfy_with;

pub use parsers::scan;

pub use parsers::skip_while;

pub use parsers::string;

pub use parsers::take;

pub use parsers::take_remainder;

pub use parsers::take_till;

pub use parsers::take_while;

pub use parsers::take_while1;

pub use parsers::token;

pub use parsers::SimpleResult;

pub use combinators::count;

pub use combinators::option;

pub use combinators::or;

pub use combinators::either;

pub use combinators::many;

pub use combinators::many1;

pub use combinators::sep_by;

pub use combinators::sep_by1;

pub use combinators::many_till;

pub use combinators::skip_many;

pub use combinators::skip_many1;

pub use combinators::matched_by;

pub use types::Buffer;

pub use types::Input;

pub use types::U8Input;

pub use types::ParseResult;

pub use either::*;

Macros

try_left
Macro for unwrapping the left side of an Either, which fails early with the opposite side. Can only be used in functions that return Either because of the early return of Right that it provides.

try_right
Dual to try_left!, see its documentation for more information.

Structs

Error
Common error for the basic Chomp parsers.

Functions

parse_only
Runs the given parser on the supplied finite input.

parse_only_str
Runs the given parser on the supplied string.