Module chomp::primitives

Module used to construct fundamental parsers and combinators.

The Primitives trait allows access to the primitive methods of the Input trait. These methods are hidden in the documentation to make it easier to read the documentation since the methods are not useful when using the library or writing primitive parsers.

Structs

Guard
This is a zero-sized type used by the Primitives trait implementation to guarantee that access to primitive methods on Input only happens when the Primitives trait has been imported.

Traits

IntoInner
Consumes self and reveals the inner state.

Primitives
Trait enabling primitive actions on an Input type.