Enum chomp::primitives::State

#[must_use]
pub enum State<'a, I: 'a, T, E> where I: 'a, T: 'a, E: 'a {
    Data(Input<'a, I>, T),
    Error(&'a [I], E),
    Incomplete(usize),
}

Primitive: Primitive inner type containing the parse-state.

Primitive

Only used by fundamental parsers and combinators.

Variants

Data

Successful parser state, first item is the input state and the second item is the contained value.

Error

Parse error state, first item is a slice from where the error occurred in the input buffer to the end of the input buffer and the second item is the error value.

Incomplete

Incomplete state, a parser attempted to request more data than available in the slice, the provided number is a guess at how many items are needed.