Struct chomp::Input

#[must_use]
pub struct Input<'a, I: 'a>(_, _);

Linear type containing the parser state, this type is threaded though bind and is also the initial type passed to a parser.

Coupled with the ParseResult type it forms the parser monad:

Fn*(Input<I>, ...) -> ParseResult<I, T, E>;

where Fn* is the appropriate closure/function trait, I the input token type (usually something like u8), ... additional parameters to the parser, T the carried type and E the potential error type.

Methods

impl<'a, I> Input<'a, I>

fn ret<T, E>(self, t: T) -> ParseResult<'a, I, T, E>

Returns t as a success value in the parsing context.

Equivalent to Haskell's return function in the Monad typeclass.

Example

use chomp::parse_only;

let r = parse_only(|i|
    // Annotate the error type
    i.ret::<_, ()>("Wohoo, success!"),
    b"some input");

assert_eq!(r, Ok("Wohoo, success!"));

fn err<T, E>(self, e: E) -> ParseResult<'a, I, T, E>

Returns e as an error value in the parsing context.

A more general version of Haskell's fail function in the Monad typeclass.

Example

use chomp::{ParseError, parse_only};

let r = parse_only(|i|
    // Annotate the value type
    i.err::<(), _>("Something went wrong"),
    b"some input");

assert_eq!(r, Err(ParseError::Error(b"some input", "Something went wrong")));

fn incomplete<T, E>(self, n: usize) -> ParseResult<'a, I, T, E>

Notifies that a parser has reached the end of the currently supplied slice but requires more data.

Primitive

Only used by fundamental parsers and combinators.

fn from_result<T, E>(self, r: Result<T, E>) -> ParseResult<'a, I, T, E>

Converts a Result into a ParseResult, preserving parser state.

To convert an Option into a ParseResult it is recommended to use Option::ok_or or Option::ok_or_else in combination with this method.

Examples

use chomp::{ParseError, parse_only};

let r = parse_only(|i| i.from_result::<_, ()>(Ok("foo")), b"test");

assert_eq!(r, Ok("foo"));

let r = parse_only(|i| i.from_result::<(), _>(Err("error message")), b"test");

assert_eq!(r, Err(ParseError::Error(&b"test"[..], "error message")));