Module chomp::types

Types which facillitates the chaining of parsers and their results.

Modules

numbering
Module containing tools for working with position aware parsers.

Structs

ParseResult
The basic return type of a parser.

Traits

Buffer
The buffers yielded parsers consuming a sequence of the input.

Input
Linear type containing the parser state, this type is threaded though bind and is also the initial type passed to a parser.

U8Input
A type alias for an Input with a token type of u8.