Trait chomp::types::numbering::Numbering

pub trait Numbering: Clone {
    type Token: Copy + PartialEq;
    fn update<B>(&mut self, _: &B)
    where
        B: Buffer<Token = Self::Token>;
    fn add(&mut self, _: Self::Token);
}

Trait for managing some kind of numbering over the parsed data.

Associated Types

type Token: Copy + PartialEq

The token type accepted by the numbering.

Required Methods

fn update<B>(&mut self, _: &B) where
B: Buffer<Token = Self::Token>,

Updates the numbering based on the contents of the buffer, adding it to the current numbering.

fn add(&mut self, _: Self::Token)

Adds the token to the numbering.